+++ src/main/java/org/cardanofoundation/rewards/entity/persistence/AggregationsInEpoch.java
@@ -2,13 +2,15 @@ package org.cardanofoundation.rewards.entity.persistence;
 
 import lombok.*;
 
+import java.math.BigInteger;
+
 @Getter
 @Setter
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
 public class AggregationsInEpoch {
-    Double sumOfFees;
-    Double sumOfWithdrawals;
-    Double sumOfDeposits;
+    BigInteger sumOfFees;
+    BigInteger sumOfWithdrawals;
+    BigInteger sumOfDeposits;
 }
