+++ src/main/java/org/cardanofoundation/rewards/repository/DbSyncWithdrawalRepository.java
@@ -3,11 +3,13 @@ package org.cardanofoundation.rewards.repository;
 import org.cardanofoundation.rewards.entity.jpa.DbSyncWithdrawal;
 import org.springframework.data.jpa.repository.Query;
 
+import java.math.BigInteger;
+
 public interface DbSyncWithdrawalRepository extends ReadOnlyRepository<DbSyncWithdrawal, Long>{
 
     @Query("""
            SELECT SUM(withdrawal.amount) from DbSyncWithdrawal AS withdrawal
              	WHERE withdrawal.transaction.block.epochNo = :epoch
             """)
-    Double getSumOfWithdrawalsInEpoch(Integer epoch);
+    BigInteger getSumOfWithdrawalsInEpoch(Integer epoch);
 }
